Move images into assets directory
