+++ site/source/Provider.tsx
@@ -7,6 +7,7 @@ import { I18nextProvider, useTranslation } from 'react-i18next'
 import { Provider as ReduxProvider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 
+import logo from '@/assets/images/logo-monentreprise.svg'
 import { DisableAnimationOnPrintProvider } from '@/components/utils/DisableAnimationContext'
 import { ThemeColorsProvider } from '@/components/utils/colors'
 import { GlobalStyle } from '@/design-system/global-style'
@@ -16,7 +17,6 @@ import { H1, H4 } from '@/design-system/typography/heading'
 import { Link } from '@/design-system/typography/link'
 import { Body, Intro } from '@/design-system/typography/paragraphs'
 import { useIframeResizer } from '@/hooks/useIframeResizer'
-import logo from '@/images/logo-monentreprise.svg'
 
 import { ServiceWorker } from './ServiceWorker'
 import { TrackingContext } from './components/ATInternetTracking'
