+++ site/source/components/simulationExplanation/logos.ts
@@ -1,0 +1,8 @@
+export { default as CIPAV } from '@/assets/images/logos-caisses-retraite/cipav.png'
+export { default as CARPIMKO } from '@/assets/images/logos-caisses-retraite/carpimko.png'
+export { default as CARMF } from '@/assets/images/logos-caisses-retraite/carmf.png'
+export { default as CARCDSF } from '@/assets/images/logos-caisses-retraite/carcdsf.jpg'
+export { default as CNBF } from '@/assets/images/logos-caisses-retraite/cnbf.png'
+export { default as CAVEC } from '@/assets/images/logos-caisses-retraite/cavec.jpg'
+export { default as IRCEC } from '@/assets/images/logos-caisses-retraite/ircec.jpg'
+export { default as CAVP } from '@/assets/images/logos-caisses-retraite/cavp.svg'
