--- site/source/images/logos-caisses-retraite/index.tsx
@@ -1,8 +1,0 @@
-export { default as CIPAV } from './cipav.png'
-export { default as CARPIMKO } from './carpimko.png'
-export { default as CARMF } from './carmf.png'
-export { default as CARCDSF } from './carcdsf.jpg'
-export { default as CNBF } from './cnbf.png'
-export { default as CAVEC } from './cavec.jpg'
-export { default as IRCEC } from './ircec.jpg'
-export { default as CAVP } from './cavp.svg'
